Move `sentry_layer` out of its module
